[feat] PopUp 간 이동 구현
