+++ Assets/Resources/UI/PopUp/SettingPopUp.prefab
@@ -3948,8 +3948,9 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2764679204670909707}
+  - component: {fileID: 1215846703468143460}
   m_Layer: 5
-  m_Name: SettingMenu
+  m_Name: SettingPopUp
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -3986,3 +3987,15 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 1920, y: 1080}
   m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &1215846703468143460
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9147359805389645435}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a86feef3a13b5ef419a264a9caf25ad3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
