+++ Assets/Resources/UI/PopUp/TitlePopUp.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 29435faf7455e62489f5c5bc39531552
+guid: ac5793d6314508a489c851da895ff6ef
 PrefabImporter:
   externalObjects: {}
   userData:
