+++ Assets/WorkSpace/LSJ/scripts/EndPopUp.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EndPopUp : BaseUI
+{
+    void Start()
+    {
+        Debug.Log("EndPopUp");
+        GetEvent("Cancellation").Click += data => Manager.UI.PopUp.ClosePopUp();
+        GetEvent("XButton").Click += data => Manager.UI.PopUp.ClosePopUp();
+    }
+
+    private void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.X))
+        {
+            Manager.UI.PopUp.ClosePopUp();
+        }
+    }
+}
